Allow updating posts

+++ app/routes/admin/$postId.tsx
@@ -1,16 +1,26 @@
 import { Post } from "@prisma/client";
 import type { ActionFunction, LoaderFunction } from "@remix-run/node";
 import { json, redirect } from "@remix-run/node";
-import { Form, useCatch, useLoaderData } from "@remix-run/react";
-import { FormEvent } from "react";
+import { Form, useActionData, useCatch, useLoaderData } from "@remix-run/react";
+import { MouseEvent, useRef } from "react";
 import invariant from "tiny-invariant";
-import Markdown from "~/components/Markdown";
-
-import { findPost, deletePost } from "~/models/post.server";
+import { deletePost, findPost, updatePost } from "~/models/post.server";
 import { requireUserId } from "~/session.server";
 
 type LoaderData = {
   post: Post;
+};
+
+const ActionTypes = {
+  Update: "update",
+  Delete: "delete",
+} as const;
+
+type ActionData = {
+  errors?: {
+    title?: string;
+    body?: string;
+  };
 };
 
 export const loader: LoaderFunction = async ({ request, params }) => {
@@ -25,16 +35,33 @@ export const loader: LoaderFunction = async ({ request, params }) => {
 };
 
 export const action: ActionFunction = async ({ request, params }) => {
-  await requireUserId(request);
+  const userId = await requireUserId(request);
   invariant(params.postId, "postId not found");
-  await deletePost(params.postId);
-  return redirect("/admin");
+
+  const formData = await request.formData();
+  const action = formData.get("_action");
+
+  if (action === ActionTypes.Delete) {
+    await deletePost(params.postId);
+    return redirect("/admin");
+  } else if (action === ActionTypes.Update) {
+    await updatePost(params.postId, {
+      userId,
+      title: formData.get("title") as string,
+      body: formData.get("body") as string,
+    });
+    return redirect(`/admin/${params.postId}`);
+  }
 };
 
 export default function AdminPostPage() {
   const data = useLoaderData() as LoaderData;
+  const actionData = useActionData() as ActionData;
 
-  const handleSubmit = (event: FormEvent) => {
+  const titleRef = useRef<HTMLInputElement>(null);
+  const bodyRef = useRef<HTMLTextAreaElement>(null);
+
+  const confirmDeletion = (event: MouseEvent<HTMLButtonElement>) => {
     const confirmed = confirm("Are you sure you want to delete this post?");
     if (!confirmed) {
       event.preventDefault();
@@ -43,18 +70,71 @@ export default function AdminPostPage() {
 
   return (
     <div>
-      <h3 className="text-2xl font-bold">{data.post.title}</h3>
-      <div>
-        <Markdown>{data.post.body}</Markdown>
-      </div>
-      <hr className="my-4" />
-      <Form method="post" onSubmit={handleSubmit}>
-        <button
-          type="submit"
-          className="rounded bg-red-500  py-2 px-4 text-white hover:bg-red-600 focus:bg-red-400"
-        >
-          Delete
-        </button>
+      <Form
+        method="post"
+        style={{
+          display: "flex",
+          flexDirection: "column",
+          gap: 8,
+          width: "100%",
+        }}
+      >
+        <div>
+          <label className="flex w-full flex-col gap-1">
+            <span>Title: </span>
+            <input
+              ref={titleRef}
+              name="title"
+              defaultValue={data.post.title}
+              className="flex-1 rounded-md border-2 border-blue-500 px-3 text-lg leading-loose"
+              aria-invalid={actionData?.errors?.title ? true : undefined}
+              aria-errormessage={
+                actionData?.errors?.title ? "title-error" : undefined
+              }
+            />
+          </label>
+          {actionData?.errors?.title && (
+            <div className="pt-1 text-red-700" id="title-error">
+              {actionData.errors.title}
+            </div>
+          )}
+        </div>
+        <div>
+          <label className="flex w-full flex-col gap-1">
+            <span>Body: </span>
+            <textarea
+              ref={bodyRef}
+              name="body"
+              rows={8}
+              defaultValue={data.post.body}
+              className="w-full flex-1 rounded-md border-2 border-blue-500 py-2 px-3 text-lg leading-6"
+              aria-invalid={actionData?.errors?.body ? true : undefined}
+              aria-errormessage={
+                actionData?.errors?.body ? "body-error" : undefined
+              }
+            />
+          </label>
+        </div>
+        <hr className="my-4" />
+        <div className="flex space-x-4">
+          <button
+            type="submit"
+            className="rounded bg-blue-500  py-2 px-4 text-white hover:bg-blue-600 focus:bg-blue-400"
+            name="_action"
+            value={ActionTypes.Update}
+          >
+            Update
+          </button>
+          <button
+            type="submit"
+            className="rounded bg-red-500  py-2 px-4 text-white hover:bg-red-600 focus:bg-red-400"
+            name="_action"
+            value={ActionTypes.Delete}
+            onClick={confirmDeletion}
+          >
+            Delete
+          </button>
+        </div>
       </Form>
     </div>
   );
